fixes1

+++ installer/sashimono-install.sh
@@ -97,11 +97,12 @@ function setup_certbot() {
     stage "Setting up letsencrypt certbot"
 
     # We assume snap is already installed (https://certbot.eff.org/instructions?ws=other&os=ubuntufocal)
-    snap install core && snap refresh core && \
-    snap install --classic certbot && \
-    ln -s /snap/bin/certbot /usr/bin/certbot || return 1
+    snap install core && snap refresh core && snap install --classic certbot
+    ! [ -f /snap/bin/certbot ] && echo "certbot not found" && return 1
+    [ -f /usr/bin/certbot ] || ln -s /snap/bin/certbot /usr/bin/certbot || return 1
 
     # Setup the certificates
+    echo "Running certbot certonly"
     certbot certonly -n -d $inetaddr --agree-tos --email sashimono@evernode.org --standalone || return 1
 
     # We need to place our script in certbook deploy hooks dir.
@@ -110,6 +111,7 @@ function setup_certbot() {
 
     # Setup deploy hook (update contract certs on certbot SSL auto-renewal)
     local deploy_hook="/etc/letsencrypt/renewal-hooks/deploy/sashimono-$inetaddr.sh"
+    echo "Setting up certbot deploy hook $deploy_hook"
     echo "#!/bin/sh
 # This script is placed by Sashimono for automatic updataing of contract SSL certs.
 # Domain name: $inetaddr
@@ -121,18 +123,18 @@ certname=$(basename \$RENEWED_LINEAGE)
 function setup_tls_certs() {
     mkdir -p $SASHIMONO_DATA/tls
 
-    if [ "$tls_cert_file" == "letsencrypt" ] ; then
+    if [ "$tls_key_file" == "letsencrypt" ] ; then
 
         ! setup_certbot && echo "Error when setting up letsencrypt SSL certificate." && rollback
 
-    elif [ "$tls_cert_file" == "-" ] ; then
+    elif [ "$tls_key_file" == "self" ] ; then
         # If user has not provided certs we generate self-signed ones.
         stage "Generating self-signed certificates"
         ! openssl req -newkey rsa:2048 -new -nodes -x509 -days 365 -keyout $SASHIMONO_DATA/contract_template/cfg/tlskey.pem \
             -out $SASHIMONO_DATA/contract_template/cfg/tlscert.pem -subj "/C=$countrycode/CN=$inetaddr" && \
             echo "Error when generating self-signed certificate." && rollback
 
-    elif [ -f "$tls_cert_file" ] && [ -f "$tls_key_file" ] ; then
+    elif [ -f "$tls_key_file" ] && [ -f "$tls_cert_file" ] ; then
 
         stage "Transfering certificate files"
